+++ Arrays/Zero_Sum_Subarray_Detector/zero_sum_subarray_detector.c
@@ -1,28 +1,27 @@
 /**
  * @file zero_sum_subarray_detector.c
- *
  * @author Vikrant A. P. (vikrant_ap@hotmail.com)
- * 
  * @date 2025-01-26
- * 
+ *
  * @copyright Copyright (c) 2025
- * 
+ *
  */
 
+#include <stddef.h>
 #include <stdio.h>
 #include <stdbool.h>
 
 /**
- * @brief Splits array into subarrays starting with the maximum 
+ * @brief Splits array into subarrays starting with the maximum
  * possible size of subarrays and checks their sum, if it is zero
  * return true else continue till subarray size reaches to one.
- * 
+ *
  * @param arr Input array
  * @param arr_size Length of the array
  * @return true if sum of the subarray elements is zero
  * @return false if subarray of sum equals to zero not found
  */
-bool zero_sum_subarray_detector(int arr[], int arr_size) {
+bool zero_sum_subarray_detector(int arr[], size_t arr_size) {
     int subarray_size = arr_size - 1;
     int sum;
 
@@ -53,10 +52,10 @@ int main() {
     //int arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9};
 
     // Calculate the size of the array
-    int arr_len = sizeof(arr) / sizeof(arr[0]);
+    size_t arr_len = sizeof(arr) / sizeof(arr[0]);
 
     bool result = zero_sum_subarray_detector(arr, arr_len);
     result ? printf("true\n") : printf("false\n");
-    
+
     return 0;
 }
